Better encapsulation of sequ
Use a setup function that will return the sequ object back to main so
that all the command line parsing and setting of sequ object defaults is
hidden from main. Added in a check for floating points. Right now based
on spec I am going to warn the user that they should use integers
instead of floats and then cast the float to an int.