Setting up end argument when using --number-lines
The user is not allowed to pass in the end arg because they do not know
the end. However, since we are now reading the entire file in first, we
can calculate the end and assign it so that --pad, --equal-width, and
--pad-spaces can be used correctly with --number-lines. That said if you
use your own --format then we use that instead.