check for --number-lines output in main program
The check for --number-lines does not need to be in every kind of
output. Instead move it out to the program and if its not there then
call your normal functions for output based on whether --format-word is
present. If it is there then call the readFile() function which has the
code necessary for reading the file in and formatting the lines numbers.